Move CtrlReset() and ResetMachineState() to Utilities.cpp.
