+++ source/Utilities.h
@@ -11,3 +11,6 @@ void InsertFloppyDisks(const UINT slot, LPSTR szImageName_drive[NUM_DRIVES], boo
 void InsertHardDisks(LPSTR szImageName_harddisk[NUM_HARDDISKS], bool& bBoot);
 void UnplugHardDiskControllerCard(void);
 void GetAppleWindowTitle();
+
+void CtrlReset();
+void ResetMachineState();
